+++ v3/newrelic/internal_test.go
@@ -42,6 +42,12 @@ var (
 		{Name: "Errors/allOther", Scope: "", Forced: true, Data: singleCount},
 		{Name: "Errors/OtherTransaction/Go/hello", Scope: "", Forced: true, Data: singleCount},
 	}, backgroundMetrics...)
+	backgroundErrorMetricsUnknownCaller = append([]internal.WantMetric{
+		{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/all", Scope: "", Forced: false, Data: nil},
+		{Name: "DurationByCaller/Unknown/Unknown/Unknown/Unknown/allOther", Scope: "", Forced: false, Data: nil},
+		{Name: "ErrorsByCaller/Unknown/Unknown/Unknown/Unknown/all", Scope: "", Forced: false, Data: nil},
+		{Name: "ErrorsByCaller/Unknown/Unknown/Unknown/Unknown/allOther", Scope: "", Forced: false, Data: nil},
+	}, backgroundErrorMetrics...)
 )
 
 type recordedLogMessage struct {
@@ -1447,7 +1453,10 @@ func TestExternalSegmentCustomFieldsWithResponse(t *testing.T) {
 	txn := app.StartTransaction("hello")
 	txn.SetWebRequestHTTP(helloRequest)
 	req, _ := http.NewRequest("GET", "https://www.something.com/path/zip/zap?secret=ssshhh", nil)
-	resp := &http.Response{Request: req}
+	resp := &http.Response{
+		Request:    req,
+		StatusCode: 13,
+	}
 	s := ExternalSegment{
 		StartTime: txn.StartSegmentNow(),
 		Response:  resp,
@@ -1476,10 +1485,11 @@ func TestExternalSegmentCustomFieldsWithResponse(t *testing.T) {
 				"component": "grpc",
 				"span.kind": "client",
 			},
-			UserAttributes:  map[string]interface{}{},
+			UserAttributes: map[string]interface{}{},
 			AgentAttributes: map[string]interface{}{
 				// "http.url" and "http.method" are not saved if
 				// library is not "http".
+				"http.statusCode": 13,
 			},
 		},
 		{